Fix import paths after edit sidebar folder restructure.
